requirements.txt
aiofiles==24.1.0
alembic==1.14.0
annotated-types==0.7.0
anyio==4.4.0
argon2-cffi==25.1.0
argon2-cffi-bindings==25.1.0
asgiref==3.9.1
Authlib==1.3.0
bcrypt==5.0.0
black==24.8.0
cachetools==5.5.2
celery==5.4.0
certifi==2025.6.15
cffi==2.0.0
chardet==5.2.0
charset-normalizer==3.4.2
click==8.2.1
colorama==0.4.6
cryptography==46.0.3
Deprecated==1.3.1
Django==5.2.4
django-cors-headers==4.7.0
djangorestframework==3.16.0
dnspython==2.8.0
ecdsa==0.19.1
email_validator==2.2.0
fastapi==0.115.2
flake8==7.1.1
google-ai-generativelanguage==0.6.15
google-api-core==2.25.1
google-api-python-client==2.172.0
google-auth==2.40.3
google-auth-httplib2==0.2.0
google-generativeai==0.8.5
googleapis-common-protos==1.70.0
greenlet==3.2.4
grpcio==1.73.0
grpcio-status==1.71.0
h11==0.16.0
httpcore==1.0.9
httplib2==0.22.0
httptools==0.6.4
httpx==0.27.2
idna==3.10
iniconfig==2.3.0
isort==5.13.2
limits==5.6.0
Mako==1.3.10
MarkupSafe==3.0.3
mccabe==0.7.0
mypy_extensions==1.1.0
packaging==25.0
passlib==1.7.4
pathspec==0.12.1
pdf2image==1.17.0
pillow==11.2.1
platformdirs==4.5.0
pluggy==1.6.0
proto-plus==1.26.1
protobuf==5.29.5
psycopg2-binary==2.9.11
pyasn1==0.6.1
pyasn1_modules==0.4.2
pycodestyle==2.12.1
pycparser==2.23
pydantic==2.9.2
pydantic_core==2.23.4
pyflakes==3.2.0
pyparsing==3.2.3
PyPDF2==3.0.1
pytest==8.3.3
python-dotenv==1.0.1
python-jose==3.3.0
python-multipart==0.0.9
PyYAML==6.0.2
rapidfuzz==3.9.3
redis==5.1.1
reportlab==4.4.1
requests==2.32.3
rsa==4.9.1
six==1.17.0
slowapi==0.1.9
sniffio==1.3.1
SQLAlchemy==2.0.36
sqlparse==0.5.3
starlette==0.40.0
tqdm==4.67.1
typing-inspection==0.4.1
typing_extensions==4.14.0
tzdata==2025.2
uritemplate==4.2.0
urllib3==2.4.0
uvicorn==0.31.0
watchfiles==1.0.5
websockets==15.0.1
wrapt==2.0.1
